requirements.txt
asgiref==3.7.2
Django==4.2.4
django-autoslug==1.9.9
django-ckeditor==6.7.0
django-cleanup==8.0.0
django-jazzmin==2.6.0
django-js-asset==2.1.0
django-taggit==4.0.0
sqlparse==0.4.4
whitenoise==6.6.0
